backend/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-JWT-Extended
flask-cors
bcrypt
psycopg2-binary
python-dotenv

